+++ vivarium/core/engine.py
@@ -10,11 +10,15 @@ import os
 import logging as log
 import pprint
 from typing import (
-    Any, Dict, Optional, Union, Tuple, Callable)
+    Any, Dict, Optional, Union, Tuple, Callable, Iterable, List, Set,
+    cast, Sequence)
 import math
 import datetime
 import time as clock
 import uuid
+import warnings
+
+import networkx as nx
 
 from vivarium.composites.toys import Proton, Electron, Sine, PoQo, ToyDivider
 from vivarium.core.store import hierarchy_depth, Store, generate_state
@@ -22,17 +26,21 @@ from vivarium.core.emitter import get_emitter
 from vivarium.core.process import (
     Process,
     ParallelProcess,
+    Step,
+    Deriver,
 )
 from vivarium.core.serialize import serialize_value
 from vivarium.core.composer import Composer
 from vivarium.library.topology import (
     delete_in,
     assoc_path,
-    inverse_topology
+    inverse_topology,
+    normalize_path,
 )
 from vivarium.library.units import units
 from vivarium.core.types import (
-    HierarchyPath, Topology, Schema, State, Update, Processes)
+    HierarchyPath, Topology, Schema, State, Update, Processes, Steps,
+    Flow)
 
 pretty = pprint.PrettyPrinter(indent=2, sort_dicts=False)
 
@@ -219,15 +227,147 @@ class InvokeProcess:
         return self.update
 
 
+class _StepGraph:
+    '''A dependency graph of :term:`steps`.
+
+    A step is just a Process object that has dependencies on other
+    steps. Unlike processes, which can be run in any order, steps run
+    every timestep. In a given timestep, each step must not run until
+    all its dependency steps have run and had their updates applied.
+
+    Note that the constructor uses any provided arguments without
+    copying them.
+
+    Attributes:
+        graph: A NetworkX DiGraph with an edge for each dependency
+            relationship and a node for each step path. If the step at
+            path ``a`` depends on the step at path ``b``, then the graph
+            will contain an edge from ``b`` to ``a``. This means that a
+            topological sort of the graph produces a valid runtime order
+            for the step. Note that the graph must be a DAG.
+        sequential_steps: A list of paths for steps that should run
+            sequentially and before the steps in ``graph``. This is
+            where we store legacy :term:`derivers` for
+            backwards-compatibility.
+    '''
+
+    def __init__(
+            self,
+            graph: Optional[nx.DiGraph] = None,
+            sequential_steps: Optional[List[HierarchyPath]] = None
+            ) -> None:
+        self._graph = graph or nx.DiGraph()
+        self._sequential_steps: List[HierarchyPath] = (
+            sequential_steps or [])
+
+    def _validate(self) -> None:
+        if not nx.is_directed_acyclic_graph(self._graph):
+            raise ValueError('Step graph must be a DAG.')
+        graph_steps = set(self._graph.nodes)
+        sequential_steps = set(self._sequential_steps)
+        intersection = graph_steps & sequential_steps
+        if intersection:
+            raise ValueError(
+                'self._graph and self._sequential_steps have '
+                f'overlapping steps: {intersection}')
+
+    def add(
+            self,
+            path: HierarchyPath,
+            dependencies: Iterable[HierarchyPath]) -> None:
+        '''Add a step to the graph.
+
+        Args:
+            path: The step object's path in the hierarchy.
+            dependencies: The hierarchy paths to each dependency.
+
+        Raises:
+            ValueError: If the graph produced by adding the step is not
+                a DAG.
+        '''
+        self._graph.add_node(path)
+        for dependency in dependencies:
+            self._graph.add_edge(dependency, path)
+        self._validate()
+
+    def add_sequential(
+            self,
+            path: HierarchyPath) -> None:
+        '''Add a step that is meant to run sequentially.
+
+        Legacy steps (:term:`derivers`) were meant to run sequentially
+        instead of being provided as a dependency graph. To support
+        these legacy steps, this method adds a step that will run after
+        all previously-added sequential steps and before any
+        non-sequential steps.
+
+        Args:
+            path: The path to the step in the hierarchy.
+        '''
+        self._sequential_steps.append(path)
+        self._validate()
+
+    def get_execution_layers(self) -> List[Set[HierarchyPath]]:
+        '''Get step execution layers, with steps represnted by paths.
+
+        An execution layer is a set of steps that can be executed in
+        parallel. The graph's execution layers are an ordered list of
+        these layers such that:
+
+        * For a given layer, every step in the layer may be executed as
+          soon as all steps in preceding layers have been executed.
+        * Every step is in as early a layer as possible.
+
+        In other words, the execution layers are the topological
+        generations of the graph, prepended by any sequential steps,
+        each in its own layer and in the order in which they were added.
+
+        Returns:
+            An ordered list of the execution layers, with each step
+            represented by its path.
+        '''
+        layers = nx.topological_generations(self._graph)
+        to_return = [set([step]) for step in self._sequential_steps]
+        to_return += [set(layer) for layer in layers]
+        return to_return
+
+    def remove(self, path: HierarchyPath) -> None:
+        '''Delete a step based on its path.
+
+        Args:
+            path: Hierarhcy path of the step to delete.
+        '''
+        if path in self._sequential_steps:
+            self._sequential_steps.remove(path)
+            return
+        to_delete = nx.algorithms.dag.descendants(self._graph, path)
+        to_delete.add(path)
+        for path_to_delete in to_delete:
+            self._graph.remove_node(path_to_delete)
+
+    def copy(self) -> '_StepGraph':
+        '''Create a copy of self.
+
+        Returns:
+            A new _StepGraph with a copy of self's graph.
+        '''
+        new = self.__class__(
+            self._graph.copy(), self._sequential_steps.copy())
+        return new
+
+
 class Engine:
     def __init__(
             self,
             processes: Optional[Processes] = None,
+            steps: Optional[Steps] = None,
+            flow: Optional[Flow] = None,
             topology: Optional[Topology] = None,
             store: Optional[Store] = None,
             initial_state: Optional[State] = None,
             experiment_id: Optional[str] = None,
             experiment_name: Optional[str] = None,
+            metadata: Optional[dict] = None,
             description: str = '',
             emitter: Union[str, dict] = 'timeseries',
             emit_topology: bool = True,
@@ -240,36 +380,46 @@ class Engine:
     ) -> None:
         """Defines simulations
 
-        Arguments:
-        * **processes** (:py:class:`dict`): A dictionary that
-        maps :term:`process` names to process objects. You will
-        usually get this from the ``processes`` attribute of the
-        dictionary from :py:meth:`vivarium.core.composer.Composer.generate`.
-        * **topology** (:py:class:`dict`): A dictionary that
-        maps process names to sub-dictionaries. These sub-dictionaries
-        map the process's port names to tuples that specify a path through
-        the :term:`tree` from the :term:`compartment` root to the
-        :term:`store` that will be passed to the process for that port.
-        * **store**: A pre-loaded Store. This is an alternative to
-        passing in processes and topology dict, which can not be
-        loaded at the same time.
-        * **initial_state** (:py:class:`dict`): By default an
-        empty dictionary, this is the initial state of the simulation.
-        * **experiment_id** (:py:class:`uuid.UUID` or :py:class:`str`):
-        A unique identifier for the experiment. A UUID will be generated
-        if none is provided.
-        * **description** (:py:class:`str`): A description of the
-        experiment. A blank string by default.
-        * **emitter** (:py:class:`dict`): An emitter configuration
-        which must conform to the specification in the documentation
-        for :py:func:`vivarium.core.emitter.get_emitter`. The experiment
-        ID will be added to the dictionary you provide as the value for
-        the key ``experiment_id``.
-        * **display_info** (:py:class:`bool`): prints experiment info
-        * **progress_bar** (:py:class:`bool`): shows a progress bar
-        * **emit_config** (:py:class:`bool`): If True, this will emit
-        the serialized processes, topology, and initial state.
-
+        Args:
+            processes: A dictionary that maps :term:`process` names to
+                process objects. You will usually get this from the
+                ``processes`` key of the dictionary from
+                :py:meth:`vivarium.core.composer.Composer.generate`.
+            steps: A dictionary that maps :term:`step` names to step
+                objects. You will usually get this from the ``steps``
+                key of the dictionary from
+                :py:meth:`vivarium.core.composer.Composer.generate`.
+            flow: A dictionary that maps :term:`step` names to sequences
+                of paths to the steps that the step depends on. You will
+                usually get this from the ``flow`` key of the dictionary
+                from
+                :py:meth:`vivarium.core.composer.Composer.generate`.
+            topology: A dictionary that maps process names to
+                sub-dictionaries. These sub-dictionaries map the
+                process's port names to tuples that specify a path
+                through the :term:`tree` from the :term:`compartment`
+                root to the :term:`store` that will be passed to the
+                process for that port.
+            store: A pre-loaded Store. This is an alternative to passing
+                in processes and topology dict, which can not be loaded
+                at the same time.
+            initial_state: By default an empty dictionary, this is the
+                initial state of the simulation.
+            experiment_id: A unique identifier for the experiment. A
+                UUID will be generated if none is provided.
+            metadata: A dictionary with additional data about the experiment,
+                which is saved by the emitter with the configuration.
+            description: A description of the experiment. A blank string
+                by default.
+            emitter: An emitter configuration which must conform to the
+                specification in the documentation for
+                :py:func:`vivarium.core.emitter.get_emitter`. The
+                experiment ID will be added to the dictionary you
+                provide as the value for the key ``experiment_id``.
+            display_info: prints experiment info
+            progress_bar: shows a progress bar
+            emit_config: If True, this will emit the serialized
+                processes, topology, and initial state.
         """
 
         self.experiment_id = experiment_id or str(uuid.uuid1())
@@ -279,17 +429,24 @@ class Engine:
         # get the processes, topology, and store
         if processes and topology and not store:
             self.processes = processes
+            self.steps = steps or {}
+            self.flow = flow or {}
             self.topology = topology
             # initialize the store
             self.state: Store = generate_state(
                 self.processes,
                 self.topology,
-                self.initial_state)
+                self.initial_state,
+                self.steps,
+                self.flow,
+            )
 
         elif store:
             self.state = store
             # get processes and topology from the store
             self.processes = self.state.get_processes()
+            self.steps = self.state.get_steps() or {}
+            self.flow = self.state.get_flow() or {}
             self.topology = self.state.get_topology()
         else:
             raise Exception(
@@ -297,6 +454,7 @@ class Engine:
 
         # display settings
         self.experiment_name = experiment_name or self.experiment_id
+        self.metadata = metadata
         self.description = description
         self.display_info = display_info
         self.progress_bar = progress_bar
@@ -310,8 +468,10 @@ class Engine:
 
         # get a mapping of all paths to processes
         self.process_paths: Dict[HierarchyPath, Process] = {}
-        self.deriver_paths: Dict[HierarchyPath, Process] = {}
-        self._find_process_paths(self.processes)
+        self._step_graph = _StepGraph()
+        self._step_paths: Dict[HierarchyPath, Process] = {}
+        self._find_process_paths(self.processes, self.flow)
+        self._find_step_paths(self.steps, self.flow)
 
         # emitter settings
         emitter_config = emitter
@@ -329,8 +489,8 @@ class Engine:
         # initialize global time
         self.experiment_time = 0.0
 
-        # run the derivers
-        self.send_updates([])
+        # run the steps
+        self._run_steps()
 
         # run the emitter
         self.emit_configuration()
@@ -346,23 +506,60 @@ class Engine:
         log.info(pf(self.topology))
 
 
+    def _add_step_path(
+            self,
+            step: Step,
+            path: HierarchyPath,
+            # None if deriver, empty list if no dependencies.
+            relative_dependencies: Optional[Sequence[HierarchyPath]],
+    ) -> None:
+        self._step_paths[path] = step
+        if relative_dependencies is None:
+            self._step_graph.add_sequential(path)
+            return
+        dependencies = [
+            path + ('..',) + dep for dep in relative_dependencies]
+        norm_dependencies = [
+            normalize_path(dep) for dep in dependencies]
+        self._step_graph.add(path, norm_dependencies)
+
     def _add_process_path(
             self,
             process: Process,
-            path: HierarchyPath
+            path: HierarchyPath,
+            flow: Flow,
     ) -> None:
-        if process.is_deriver():
-            self.deriver_paths[path] = process
+        if process.is_step():
+            warnings.warn(
+                f'Found a step {path} in the processes dict. This is '
+                'deprecated. Steps should be specified in the steps '
+                'dict instead.',
+                category=FutureWarning,
+            )
+            step = cast(Step, process)
+            name = path[-1]
+            self._add_step_path(step, path, flow.get(name))
         else:
             self.process_paths[path] = process
 
     def _find_process_paths(
             self,
-            processes: Processes
+            processes: Processes,
+            flow: Flow,
     ) -> None:
         tree = hierarchy_depth(processes)
         for path, process in tree.items():
-            self._add_process_path(process, path)
+            self._add_process_path(process, path, flow)
+
+    def _find_step_paths(
+            self,
+            steps: Steps,
+            flow: Flow,
+    ) -> None:
+        tree = hierarchy_depth(steps)
+        for path, step in tree.items():
+            name = path[-1]
+            self._add_step_path(step, path, flow.get(name))
 
     def emit_configuration(self) -> None:
         """Emit experiment configuration."""
@@ -371,6 +568,7 @@ class Engine:
             'experiment_id': self.experiment_id,
             'name': self.experiment_name,
             'description': self.description,
+            'metadata': self.metadata,
             'topology': self.topology
             if self.emit_topology else None,
             'processes': serialize_value(self.processes)
@@ -542,8 +740,12 @@ class Engine:
         if not update:
             return
 
-        topology_updates, process_updates, deletions = self.state.apply_update(
-            update, state)
+        (
+            topology_updates, process_updates, step_updates,
+            flow_updates, deletions
+        ) = self.state.apply_update(update, state)
+
+        flow_update_dict = dict(flow_updates)
 
         if topology_updates:
             for path, topology_update in topology_updates:
@@ -552,7 +754,14 @@ class Engine:
         if process_updates:
             for path, process in process_updates:
                 assoc_path(self.processes, path, process)
-                self._add_process_path(process, path)
+                self._add_process_path(process, path, {})
+
+        if step_updates:
+            for path, step in step_updates:
+                dependencies = flow_update_dict.get(path)
+                assoc_path(self.steps, path, step)
+                self._add_step_path(step, path, dependencies)
+
 
         if deletions:
             for deletion in deletions:
@@ -574,32 +783,43 @@ class Engine:
             if starts_with(path, deletion):
                 del self.process_paths[path]
 
-        for path in list(self.deriver_paths.keys()):
+        for path in list(self._step_paths):
             if starts_with(path, deletion):
-                del self.deriver_paths[path]
+                try:
+                    self._step_graph.remove(path)
+                except nx.exception.NetworkXError as e:
+                    # The step might have been deleted already.
+                    msg = f'The node {path} is not in the graph.'
+                    if e.args[0] != msg:
+                        raise e
+                del self._step_paths[path]
 
-    def run_derivers(self) -> None:
-        """Run all the derivers in the simulation."""
-        paths = list(self.deriver_paths.keys())
-        for path in paths:
-            # deriver could have been deleted by another deriver
-            deriver = self.deriver_paths.get(path)
-            if deriver:
-                # timestep shouldn't influence derivers
+    def _run_steps(self) -> None:
+        '''Run all the steps in the simulation.'''
+        layers = self._step_graph.get_execution_layers()
+        for layer in layers:
+            deferred_updates: List[Tuple[Defer, Store]] = []
+            for path in layer:
+                step = self._step_paths.get(path)
+                if not step:
+                    # Step was deleted by a previous step.
+                    continue
+                # Timestep shouldn't influence steps.
                 # TODO(jerry): Do something cleaner than having
                 #  generate_paths() add a schema attribute to the Deriver.
                 #  PyCharm's type check reports:
                 #    Type Process doesn't have expected attribute 'schema'
                 update, store = self.calculate_update(
-                    path, deriver, 0)
+                    path, step, 0)
+                deferred_updates.append((update, store))
+            for update, store in deferred_updates:
                 self.apply_update(update.get(), store)
-
 
     def send_updates(
             self,
             update_tuples: list
     ) -> None:
-        """Apply updates and run derivers.
+        """Apply updates and run steps.
 
         Args:
             update_tuples: List of tuples ``(update, state)`` where
@@ -610,7 +830,7 @@ class Engine:
             update, state = update_tuple
             self.apply_update(update.get(), state)
 
-        self.run_derivers()
+        self._run_steps()
 
     def update(
             self,
@@ -635,7 +855,8 @@ class Engine:
             full_step = math.inf
 
             # find any parallel processes that were removed and terminate them
-            for terminated in self.parallel.keys() - self.process_paths.keys():
+            for terminated in self.parallel.keys() - (
+                    self.process_paths.keys() | self._step_paths.keys()):
                 self.parallel[terminated].end()
                 del self.parallel[terminated]
 
@@ -1307,20 +1528,248 @@ def test_custom_divider() -> None:
     composer = ToyDivider({
         'agent_id': agent_id,
         'divider': {
-            'x_default': 3,
-            'x_division_threshold': 25,
+            'x_division_threshold': 3,
         }
     })
     composite = composer.generate(path=('agents', agent_id))
 
     experiment = Engine(
         processes=composite.processes,
+        steps=composite.steps,
+        flow=composite.flow,
         topology=composite.topology,
     )
 
-    experiment.update(80)
+    experiment.update(8)
     data = experiment.emitter.get_data()
-    print(pf(data))
+
+    expected_data = {
+        0.0: {'agents': {'1': {'variable': {'x': 0, '2x': 0}}}},
+        1.0: {'agents': {'1': {'variable': {'x': 1, '2x': 2}}}},
+        2.0: {'agents': {'1': {'variable': {'x': 2, '2x': 4}}}},
+        3.0: {'agents': {'1': {'variable': {'x': 3, '2x': 6}}}},
+        4.0: {'agents': {'1': {'variable': {'x': 4, '2x': 8}}}},
+        5.0: { 'agents': { '10': {'variable': {'x': 2.0, '2x': 4.0}},
+                           '11': {'variable': {'x': 2.0, '2x': 4.0}}}},
+        6.0: { 'agents': { '10': {'variable': {'x': 3.0, '2x': 6.0}},
+                           '11': {'variable': {'x': 3.0, '2x': 6.0}}}},
+        7.0: { 'agents': { '10': {'variable': {'x': 4.0, '2x': 8.0}},
+                           '11': {'variable': {'x': 4.0, '2x': 8.0}}}},
+        8.0: { 'agents': { '100': {'variable': {'x': 2.0, '2x': 4.0}},
+                           '101': {'variable': {'x': 2.0, '2x': 4.0}},
+                           '110': {'variable': {'x': 2.0, '2x': 4.0}},
+                           '111': {'variable': {'x': 2.0, '2x': 4.0}}}}
+    }
+    assert data == expected_data
+
+
+
+class TestStepGraph:
+
+    @staticmethod
+    def test_step_graph_execution_layers() -> None:
+        tg = _StepGraph()
+        tg.add(('a',), [])
+        tg.add(('b',), [])
+        tg.add(('c',), [('a',), ('b',)])
+        tg.add(('d',), [('c',)])
+
+        layers = list(tg.get_execution_layers())
+        expected_layers = [
+            {('a',), ('b',)},
+            {('c',)},
+            {('d',)},
+        ]
+        assert layers == expected_layers
+
+    @staticmethod
+    def test_step_graph_sequential() -> None:
+        tg = _StepGraph()
+        tg.add(('a',), [])
+        tg.add(('b',), [])
+        tg.add_sequential(('c',))
+
+        layers = list(tg.get_execution_layers())
+        expected_layers = [
+            {('c',)},
+            {('a',), ('b',)},
+        ]
+        assert layers == expected_layers
+
+    @staticmethod
+    def test_step_graph_removal() -> None:
+        tg = _StepGraph()
+        tg.add(('a',), [])
+        tg.add(('b',), [('a',)])
+        tg.add(('c',), [('b',)])
+
+        tg.remove(('a',))
+        layers = list(tg.get_execution_layers())
+
+        assert layers == []
+
+
+def test_runtime_order() -> None:
+
+    class RuntimeOrderProcess(Process):
+
+        def ports_schema(self) -> Schema:
+            return {
+                'store': {
+                    'var': {
+                        '_default': 0
+                    }
+                }
+            }
+
+        def next_update(self, timestep: float, states: State) -> Update:
+            _ = states
+            self.parameters['execution_log'].append(self.name)
+            return {}
+
+
+    class RuntimeOrderStep(Step):
+
+        def ports_schema(self) -> Schema:
+            return {
+                'store': {
+                    'var': {
+                        '_default': 0
+                    }
+                }
+            }
+
+        def next_update(self, timestep: float, states: State) -> Update:
+            _ = states
+            self.parameters['execution_log'].append(self.name)
+            return {}
+
+    class RuntimeOrderDeriver(Deriver):
+
+        def ports_schema(self) -> Schema:
+            return {
+                'store': {
+                    'var': {
+                        '_default': 0
+                    }
+                }
+            }
+
+        def next_update(self, timestep: float, states: State) -> Update:
+            _ = states
+            self.parameters['execution_log'].append(self.name)
+            return {}
+
+    class RuntimeOrderComposer(Composer):
+
+        def generate_processes(
+                self, config: Optional[dict]) -> Dict[str, Any]:
+            config = cast(dict, config or {})
+            proc1 = RuntimeOrderProcess({
+                'name': 'process1',
+                'time_step': 1,
+                'execution_log': config['execution_log'],
+            })
+            proc2 = RuntimeOrderProcess({
+                'name': 'process2',
+                'time_step': 2,
+                'execution_log': config['execution_log'],
+            })
+            deriver = RuntimeOrderDeriver({
+                'name': 'deriver',
+                'execution_log': config['execution_log'],
+            })
+            return {
+                'p1': proc1,
+                'p2': proc2,
+                'd': deriver,
+            }
+
+        def generate_steps(self, config: Optional[dict]) -> Steps:
+            config = config or {}
+            step1 = RuntimeOrderStep({
+                'name': 'step1',
+                'execution_log': config['execution_log'],
+            })
+            step2 = RuntimeOrderStep({
+                'name': 'step2',
+                'execution_log': config['execution_log'],
+            })
+            step3 = RuntimeOrderStep({
+                'name': 'step3',
+                'execution_log': config['execution_log'],
+            })
+            return {
+                's1': step1,
+                's2': step2,
+                's3': step3,
+            }
+
+        def generate_flow(self, config: Optional[dict]) -> Steps:
+            config = config or {}
+            return {
+                's1': [],
+                's2': [('s1',)],
+                's3': [('s1',)],
+            }
+
+
+        def generate_topology(self, config: Optional[dict]) -> Topology:
+            return {
+                'p1': {
+                    'store': ('store',),
+                },
+                'p2': {
+                    'store': ('store',),
+                },
+                'd': {
+                    'store': ('store',),
+                },
+                's1': {
+                    'store': ('store',),
+                },
+                's2': {
+                    'store': ('store',),
+                },
+                's3': {
+                    'store': ('store',),
+                },
+            }
+
+    execution_log: List[str] = []
+    composer = RuntimeOrderComposer()
+    composite = composer.generate({'execution_log': execution_log})
+    experiment = Engine(
+        processes=composite.processes,
+        steps=composite.steps,
+        flow=composite.flow,
+        topology=composite.topology,
+    )
+    experiment.update(4)
+    expected_log = [
+        ('deriver', 'step1'),
+        {'step2', 'step3'},
+        {'process1', 'process2'},
+        ('deriver', 'step1'),
+        {'step2', 'step3'},
+        {'process1'},
+        ('deriver', 'step1'),
+        {'step2', 'step3'},
+        {'process1', 'process2'},
+        ('deriver', 'step1'),
+        {'step2', 'step3'},
+        {'process1'},
+        ('deriver', 'step1'),
+        {'step2', 'step3'},
+    ]
+    for expected_group in expected_log:
+        num = len(expected_group)
+        group = execution_log[0:num]
+        execution_log = execution_log[num:]
+        if isinstance(expected_group, tuple):
+            assert tuple(group) == expected_group
+        elif isinstance(expected_group, set):
+            assert set(group) == expected_group
 
 
 if __name__ == '__main__':
@@ -1335,3 +1784,4 @@ if __name__ == '__main__':
     # test_2_store_1_port()
     # test_units()
     test_custom_divider()
+    # test_runtime_order()